Extract the messageDispatcher class from SQQMessageListener as discussed on friday
